requirements.txt
chromadb>=0.4.0
click>=8.1.0
rich>=13.0.0
questionary>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.0.0
